Added SignupForm and SignupDirectory pages and removed redundant form components
